preparring to delete columbia b/c it doesnt do anything
