+++ Dockerfile.testing.Dockerfile
@@ -44,11 +44,6 @@ COPY vendor $GOPATH/src/github.com/Cray-HPE/hms-reds/vendor
 FROM base
 
 CMD ["sh", "-c", "go test -cover -v github.com/Cray-HPE/hms-reds/cmd/reds; \
-    go test -cover -v github.com/Cray-HPE/hms-reds/internal/snmp/common; \
-    go test -cover -v github.com/Cray-HPE/hms-reds/internal/snmp/common; \
-    go test -cover -v github.com/Cray-HPE/hms-reds/internal/snmp/dell; \
-    go test -cover -v github.com/Cray-HPE/hms-reds/internal/storage; \
-    go test -cover -v github.com/Cray-HPE/hms-reds/internal/storage/etcd; \
     go test -cover -v github.com/Cray-HPE/hms-reds/internal/mapping; \
     go test -cover -v github.com/Cray-HPE/hms-reds/internal/model"]
