--- internal/storage/mock/kvmock.go
@@ -1,54 +1,0 @@
-// MIT License
-//
-// (C) Copyright [2019, 2021] Hewlett Packard Enterprise Development LP
-//
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the "Software"),
-// to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
-// OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
-// ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-// OTHER DEALINGS IN THE SOFTWARE.
-
-package mock
-
-import (
-	"github.com/mitchellh/mapstructure"
-)
-
-type KvMock struct {
-	storage map[string]interface{}
-}
-
-func NewKvMock() *KvMock {
-	var s KvMock
-	s.storage = make(map[string]interface{})
-	return &s
-}
-
-func (kv KvMock) Store(key string, value interface{}) error {
-	kv.storage[key] = value
-	return nil
-}
-func (kv KvMock) Lookup(key string, output interface{}) error {
-	value := kv.storage[key]
-	err := mapstructure.Decode(value, output)
-	return err
-}
-func (kv KvMock) Delete(key string) error {
-	delete(kv.storage, key)
-	return nil
-}
-func (kv KvMock) LookupKeys(keyPath string) (keys []string, err error) {
-	return
-}
